refactor: Consolidate terminal launch and clipboard logic into lib
